+++ src/index.ts
@@ -2,7 +2,7 @@ import { registry } from './registry.js'
 import { SimoneError } from './errors.js'
 import type { MockModuleInstance } from './types.js'
 
-export type { MockedModule, MockModuleInstance, Expectation, ExpectationWithArgs, FunctionKeys } from './types.js'
+export type { MockedModule, MockModuleInstance, Expectation, ExpectationWithArgs, SyncExpectationWithArgs, AsyncExpectationWithArgs, FunctionKeys } from './types.js'
 export { SimoneError, SimoneAlreadyFailedError } from './errors.js'
 
 export function mockModule<Module>(module: Promise<Module>): MockModuleInstance<Module> {