+++ src/types.ts
@@ -10,11 +10,20 @@ export interface Expectation<Fn> {
   withArgs(...args: Fn extends (...a: infer A) => any ? A : never): ExpectationWithArgs<Fn>
 }
 
-export interface ExpectationWithArgs<Fn> {
+export type ExpectationWithArgs<Fn> =
+  Fn extends (...a: any[]) => Promise<any>
+    ? AsyncExpectationWithArgs<Fn>
+    : SyncExpectationWithArgs<Fn>
+
+export interface SyncExpectationWithArgs<Fn> {
   returns(value: Fn extends (...a: any[]) => infer R ? R : never): void
   throws(error: unknown): void
+  calls(fn: Fn): void
+}
+
+export interface AsyncExpectationWithArgs<Fn> {
   resolves(value: Fn extends (...a: any[]) => Promise<infer V> ? V : never): void
-  rejects(error: Fn extends (...a: any[]) => Promise<any> ? unknown : never): void
+  rejects(error: unknown): void
   calls(fn: Fn): void
 }